refactor selectors to utilities
